DABOM-318 feat : UsageRecordService 분리
